+++ Ticket.Presentation/ViewModels/BillingReportWithProject.cs
@@ -8,8 +8,8 @@ namespace Ticket.Presentation.ViewModels
         public string NameSurname { get; set; }
         public string Description { get; set; }
         public string ProductName { get; set; }
-        public int TotalWorkDay { get; set; }
-        public int TotalWorkHour { get; set; }
+        public double TotalWorkDay { get; set; }
+        public double TotalWorkHour { get; set; }
         public double ManHour { get; set; }
         public double TotalPrice { get; set; }
         public string ProjectName { get; set; }
