+++ Ticket.Presentation/ViewModels/EmployeePerformanceReport.cs
@@ -6,8 +6,8 @@ namespace Ticket.Presentation.ViewModels
         public string TicketDescription { get; set; }
         public string SubTicketDescription { get; set; }
         public string NameSurname { get; set; }
-        public int TotalWorkHour { get; set; }
-        public int TotalWorkDay { get; set; }
+        public double TotalWorkHour { get; set; }
+        public double TotalWorkDay { get; set; }
         public DateTime StartDate { get; set; }
         public DateTime EndDate { get; set; }
         public string CardName { get; set; }
